requirements.txt
flask
scapy
requests
python-dotenv
pyyaml
beautifulsoup4
